+++ src/commands/ezcopy.rs
@@ -18,11 +18,30 @@ use crate::img::tracks::Method;
 use crate::{DYNERR,STDRESULT};
 use crate::lang::{applesoft,integer,merlin,is_lang};
 
+struct Source {
+    pub fimg: FileImage,
+    pub fused_path: String
+}
+
 enum Destination {
     /// disk object, path to image, path inside image
     Dimg(Box<dyn DiskFS>,String,String),
     /// path on the host system
     Host(String)
+}
+
+/// Use native path module to extract filename and error out if not UTF8,
+/// we expect this should work for either host or image paths, including on Windows.
+fn extract_ordinary_filename(path: &str) -> Result<String,DYNERR> {
+    let pbuf = PathBuf::from(path);
+    let Some(os_fname) = pbuf.file_name() else {
+        log::error!("could not extract filename from {}",path);
+        return Err(Box::new(CommandError::FileNotFound));
+    };
+    let Some(fname) = os_fname.to_str() else {
+        return Err(Box::new(CommandError::UnsupportedFormat));
+    };
+    Ok(fname.to_string())
 }
 
 /// Parse a path the starts with a disk image and ends with a path inside the disk image.
@@ -50,20 +69,43 @@ fn parse_fused_path(fused: &str,dimg_patt: &Regex) -> Result<(String,String),DYN
 /// Combine src_path and dst_path using logic that the user likely expects.
 /// If there is 1 source and the destination is not null, and not an existing directory,
 /// then the destination is used as is.  Otherwise the source filename is joined to the destination path.
-fn revise_destination_path(src_path: &str, dst_path: &str, src_count: usize, dst_dir_exists: bool) -> Result<String,DYNERR> {
-    let new_path = match (src_count,dst_dir_exists,dst_path.is_empty()) {
-        (1,false,false) => PathBuf::from(dst_path),
+fn revise_destination_path(src_path: &str, dst_path: &str, src_count: usize, dst_dir_exists: bool, dst_is_img: bool) -> Result<String,DYNERR> {
+    match (src_count,dst_dir_exists,dst_path.is_empty()) {
+        (1,false,false) => Ok(dst_path.to_owned()),
         _ => {
-            match PathBuf::from(src_path).file_name() {
-                Some(fname) => PathBuf::from(dst_path).join(fname),
-                None => return Err(Box::new(CommandError::FileNotFound))
+            let fname = extract_ordinary_filename(src_path)?;
+            match dst_is_img {
+                true => match dst_path.len() {
+                    0 => Ok(fname.to_owned()),
+                    _ => Ok([dst_path,"/",fname.as_str()].concat())
+                },
+                false => match PathBuf::from(dst_path).join(fname).as_os_str().to_str() {
+                    Some(s) => Ok(s.to_owned()),
+                    None => Err(Box::new(CommandError::UnsupportedFormat))
+                }
             }
         }
-    };
-    match new_path.as_os_str().to_str() {
-        Some(s) => Ok(s.to_owned()),
-        None => Err(Box::new(CommandError::FileNotFound))
     }
+}
+
+/// In some cases we are given a new filename in the destination path, and that has to be detected
+/// at the time when a file image is created, i.e., during the gather step, because when the
+/// file image is created an invalid filename produces an error.  We also will automatically
+/// strip any file extensions that are specified in `strip` which can help avoid length violations.
+/// It is OK and necessary to later call `revise_destination_path` to handle an implied join.
+fn revise_fimg_path(src_path: &str, dst_path: &str, src_count: usize, dst_dir_exists: bool, strip: Vec<&str>) -> String {
+    let mut host_path = match (src_count, dst_dir_exists, dst_path.is_empty()) {
+        (1,false,false) => dst_path.to_string(),
+        _ => src_path.to_string()
+    };
+    let l = host_path.len();
+    for x in strip {
+        if l > x.len() && host_path.to_lowercase().ends_with(x) {
+            host_path.truncate(l-x.len());
+            return host_path;
+        }
+    }
+    host_path
 }
 
 /// Pack up data into a file image after transforming to the emulated system's format.
@@ -102,7 +144,7 @@ fn smart_pack(fimg: &mut FileImage, dat: &[u8], load_addr: Option<usize>) -> STD
     }
 }
 
-/// Unpack a file image and transform the data to a string that is readable and invertible on the host system.
+/// Unpack a file image and possibly transform the data to a string that is readable and invertible on the host system.
 /// In this direction there is a chance file system hints can be used for identification, while in some
 /// cases it is still necessary to parse the whole slice.
 fn smart_unpack(fimg: &FileImage) -> Result<UnpackedData,DYNERR> {
@@ -153,45 +195,52 @@ fn smart_unpack(fimg: &FileImage) -> Result<UnpackedData,DYNERR> {
     }
 }
 
-fn gather(src: Vec<String>,dst: &Destination,dimg_patt: &Regex,cmd: &clap::ArgMatches) -> Result<Vec<FileImage>,DYNERR> {
+fn gather(src: Vec<String>,dst: &Destination,dst_dir_exists: bool,dimg_patt: &Regex,cmd: &clap::ArgMatches) -> Result<Vec<Source>,DYNERR> {
     let mut ans = Vec::new();
     let fmt = super::get_fmt(cmd)?;
     let load_addr: Option<usize> = match cmd.get_one::<String>("addr") {
         Some(a) => Some(usize::from_str(a)?),
         _ => None
     };
+    let src_count = src.len();
 
-    for src_path in src {
+    for fused_path in src {
 
-        match dimg_patt.is_match(&src_path) {
+        match dimg_patt.is_match(&fused_path) {
             true => {
-                let (path_to,path_in) = parse_fused_path(&src_path,dimg_patt)?;
+                let (path_to,path_in) = parse_fused_path(&fused_path,dimg_patt)?;
                 let mut src_disk = crate::create_fs_from_file(&path_to,fmt.as_ref())?;
                 src_disk.get_img().change_method(Method::from_str(cmd.get_one::<String>("method").unwrap())?);
                 match src_disk.glob(&path_in,false) {
                     Ok(vlist) => {
+                        if vlist.len() == 0 {
+                            log::error!("no matches to source path {}",path_in);
+                            return Err(Box::new(CommandError::FileNotFound));
+                        }
                         for v in vlist {
-                            ans.push(src_disk.get(&v)?);
+                            ans.push(Source {fimg: src_disk.get(&v)?, fused_path: [path_to.as_str(),v.as_str()].concat()});
                         }
                     },
-                    Err(_) => ans.push(src_disk.get(&path_in)?)
+                    Err(_) => ans.push(Source {fimg: src_disk.get(&path_in)?, fused_path})
                 }
             },
             false => {
-                match (dst,std::fs::read(&src_path)) {
+                match (dst,std::fs::read(&fused_path)) {
                     (Destination::Host(_),_) => {
                         log::error!("refusing host-to-host copy");
                         return Err(Box::new(CommandError::InvalidCommand))
                     },
-                    (Destination::Dimg(dst_disk,_,_),Ok(dat)) => {
-                        let pbuf = PathBuf::from(src_path);
-                        let filename = match pbuf.file_name() {
-                            Some(p) => p.to_string_lossy(),
-                            None => return Err(Box::new(CommandError::FileNotFound))
+                    (Destination::Dimg(dst_disk,_,raw_dst_path),Ok(dat)) => {
+                        let fname = extract_ordinary_filename(&fused_path)?;
+                        let dummy = dst_disk.new_fimg(None,true,"dummy")?;
+                        let strip = match dummy.file_system.as_str() {
+                            "prodos" | "a2 dos" => vec![".json",".txt",".bas",".abas",".ibas"],
+                            _ => vec![".json"]
                         };
-                        let mut fimg = dst_disk.new_fimg(None,true,&filename.to_string())?;
+                        let dst_path = revise_fimg_path(&fname,raw_dst_path,src_count,dst_dir_exists,strip);
+                        let mut fimg = dst_disk.new_fimg(None,true,&dst_path)?;
                         smart_pack(&mut fimg,&dat,load_addr)?;
-                        ans.push(fimg);
+                        ans.push(Source {fimg,fused_path});
                     },
                     (_,Err(e)) => return Err(Box::new(e))
                 }
@@ -226,8 +275,8 @@ pub fn ezcopy(cmd: &clap::ArgMatches) -> STDRESULT {
         }
     };
     let dst_dir_exists = match &mut dst {
-        Destination::Dimg(disk, _, path_inside) => {
-            log::info!("image destination {}",path_inside);
+        Destination::Dimg(disk, path_to, path_inside) => {
+            log::info!("image destination {}",path_to);
             match disk.catalog_to_vec(path_inside) {
                 Ok(_) => true,
                 Err(_) => false
@@ -242,7 +291,7 @@ pub fn ezcopy(cmd: &clap::ArgMatches) -> STDRESULT {
             PathBuf::from(target_path.as_str()).is_dir()
         }
     };
-    let mut src_list = gather(path_list,&dst,&dimg_patt,cmd)?;
+    let mut src_list = gather(path_list,&dst,dst_dir_exists,&dimg_patt,cmd)?;
 
     // Second stage, write to destination
 
@@ -250,18 +299,18 @@ pub fn ezcopy(cmd: &clap::ArgMatches) -> STDRESULT {
     for src in &mut src_list {
         match &mut dst {
             Destination::Dimg(dst_disk, _, raw_dst_path) => {
-                let dst_path = revise_destination_path(&src.full_path, &raw_dst_path, src_count, dst_dir_exists)?;
-                log::info!("copy {} -> {}",src.full_path,dst_path);
-                dst_disk.put_at(&dst_path,src)?;
+                let dst_path = revise_destination_path(&src.fimg.full_path, &raw_dst_path, src_count, dst_dir_exists, true)?;
+                log::info!("copy {} -> {}",src.fused_path,dst_path);
+                dst_disk.put_at(&dst_path,&mut src.fimg)?;
             },
             Destination::Host(raw_dst_path) => {
-                let dst_path = revise_destination_path(&src.full_path, raw_dst_path, src_count, dst_dir_exists)?;
+                let dst_path = revise_destination_path(&src.fimg.full_path, raw_dst_path, src_count, dst_dir_exists, false)?;
                 if PathBuf::from(dst_path.as_str()).is_file() {
                     log::error!("destination {} already exists as a file",dst_path);
                     return Err(Box::new(CommandError::InvalidCommand));
                 }
-                log::info!("copy {} -> {}",src.full_path,dst_path);
-                match smart_unpack(src)? {
+                log::info!("copy {} -> {}",src.fused_path,dst_path);
+                match smart_unpack(&src.fimg)? {
                     UnpackedData::Binary(dat) => std::fs::write(&dst_path,&dat).expect("host file system error"),
                     UnpackedData::Text(s) => std::fs::write(&dst_path,s.as_bytes()).expect("host file system error"),
                     UnpackedData::Records(r) => {