+++ tests/cli_test_meta.rs
@@ -1,7 +1,5 @@
-use assert_cmd::prelude::*; // Add methods on commands
+use assert_cmd::cargo; // Add methods on commands
 use predicates::prelude::*; // Used for writing assertions
-use std::process::{Command,Stdio}; // Run programs
-use std::io::Write;
 use tempfile;
 use json;
 type STDRESULT = Result<(),Box<dyn std::error::Error>>;
@@ -106,7 +104,7 @@ const WOZ2_EXPECTED_FILTERED: &str = "
 #[test]
 fn get_meta_do() -> STDRESULT {
     let expected = json::stringify_pretty(json::parse("{\"do\":{}}").expect("json parsing failed"),4);
-    let mut cmd = Command::cargo_bin("a2kit")?;
+    let mut cmd = cargo::cargo_bin_cmd!("a2kit");
     let dir = tempfile::tempdir()?;
     let dimg_path = dir.path().join("dos33.do");
     // first make disk
@@ -116,7 +114,7 @@ fn get_meta_do() -> STDRESULT {
         .assert()
         .success();
     // check the metadata
-    cmd = Command::cargo_bin("a2kit")?;
+    cmd = cargo::cargo_bin_cmd!("a2kit");
     cmd.arg("get")
         .arg("-t").arg("meta").arg("-d").arg(&dimg_path)
         .assert()
@@ -127,7 +125,7 @@ fn get_meta_do() -> STDRESULT {
 
 #[test]
 fn put_get_meta_woz2() -> STDRESULT {
-    let mut cmd = Command::cargo_bin("a2kit")?;
+    let mut cmd = cargo::cargo_bin_cmd!("a2kit");
     let dir = tempfile::tempdir()?;
     let dimg_path = dir.path().join("woz2.woz");
     // first make disk
@@ -138,26 +136,21 @@ fn put_get_meta_woz2() -> STDRESULT {
         .success();
 
     // set items in the INFO and META chunks
-    cmd = Command::cargo_bin("a2kit")?;
-    let mut child = cmd.arg("put")
+    cmd = cargo::cargo_bin_cmd!("a2kit");
+    cmd.arg("put")
         .arg("-t").arg("meta").arg("-d").arg(&dimg_path)
-        .stdin(Stdio::piped())
-        .spawn()
-        .expect("failed to spawn child process");
-    let mut stdin = child.stdin.take().expect("Failed to open stdin");
-    std::thread::spawn(move || {
-        stdin.write_all(WOZ2_PUT_ITEMS.as_bytes()).expect("Failed to write to stdin");
-    });
-    child.wait_with_output().expect("Failed to read stdout");
+        .write_stdin(WOZ2_PUT_ITEMS.as_bytes())
+        .assert()
+        .success();
     
     // check the metadata
-    cmd = Command::cargo_bin("a2kit")?;
-    let child = cmd.arg("get")
+    cmd = cargo::cargo_bin_cmd!("a2kit");
+    let output = cmd.arg("get")
         .arg("-t").arg("meta").arg("-d").arg(&dimg_path)
-        .stdout(Stdio::piped())
-        .spawn()
-        .expect("failed to spawn child process");    
-    let output = child.wait_with_output().expect("Failed to read stdout");
+        .assert()
+        .success()
+        .get_output().clone();
+
     // take string to object and back to get the format consistent
     let woz2_exp = WOZ2_EXPECTED.replace("2.2.0",PK_VERS);
     let expected = json::stringify_pretty(json::parse(&woz2_exp).expect("json parsing failed"),4);
@@ -168,7 +161,7 @@ fn put_get_meta_woz2() -> STDRESULT {
 
 #[test]
 fn put_get_meta_woz2_filtered() -> STDRESULT {
-    let mut cmd = Command::cargo_bin("a2kit")?;
+    let mut cmd = cargo::cargo_bin_cmd!("a2kit");
     let dir = tempfile::tempdir()?;
     let dimg_path = dir.path().join("woz2.woz");
     // first make disk
@@ -179,27 +172,22 @@ fn put_get_meta_woz2_filtered() -> STDRESULT {
         .success();
 
     // set items in the INFO and META chunks, but use filter to pass only META
-    cmd = Command::cargo_bin("a2kit")?;
-    let mut child = cmd.arg("put")
+    cmd = cargo::cargo_bin_cmd!("a2kit");
+    cmd.arg("put")
         .arg("-t").arg("meta").arg("-d").arg(&dimg_path)
         .arg("-f").arg("/woz2/meta/")
-        .stdin(Stdio::piped())
-        .spawn()
-        .expect("failed to spawn child process");
-    let mut stdin = child.stdin.take().expect("Failed to open stdin");
-    std::thread::spawn(move || {
-        stdin.write_all(WOZ2_PUT_ITEMS.as_bytes()).expect("Failed to write to stdin");
-    });
-    child.wait_with_output().expect("Failed to read stdout");
+        .write_stdin(WOZ2_PUT_ITEMS.as_bytes())
+        .assert()
+        .success();
     
     // check the metadata
-    cmd = Command::cargo_bin("a2kit")?;
-    let child = cmd.arg("get")
+    cmd = cargo::cargo_bin_cmd!("a2kit");
+    let output = cmd.arg("get")
         .arg("-t").arg("meta").arg("-d").arg(&dimg_path)
-        .stdout(Stdio::piped())
-        .spawn()
-        .expect("failed to spawn child process");    
-    let output = child.wait_with_output().expect("Failed to read stdout");
+        .assert()
+        .success()
+        .get_output().clone();
+
     // take string to object and back to get the format consistent
     let woz2_exp = WOZ2_EXPECTED_FILTERED.replace("2.2.0",PK_VERS);
     let expected = json::stringify_pretty(json::parse(&woz2_exp).expect("json parsing failed"),4);