+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.1.2] - 2026-01-17
+
+### Fixes
+
+* patches to smart copy
+    - fix some issues with paths and error handling
+    - make the language detection a little smarter
+* add tests and update test code
+
 ## [4.1.1] - 2026-01-11
 
 ### Fixes